표기 리팩토링

+++ AcousticWave/include/HornFDTD.h
@@ -7,8 +7,8 @@ private: /* parameters */
 	double decay = 1.0;
 
 public: /* constructor */
-	HornFDTD(double _length, double _period, unsigned long long _numt);
-	HornFDTD(double _length, double _period, unsigned long long _numt, double _sonic, double _decay);
+	HornFDTD(double _length, double _period, size_t _numt);
+	HornFDTD(double _length, double _period, size_t _numt, double _sonic, double _decay);
 
 protected: /* virtual */
 	bool solve() override;
