+++ AcousticWave/include/PipeFDTD.h
@@ -7,8 +7,8 @@ private: /* parameters */
 	double decay = 1.0;
 
 public: /* constructor */
-	PipeFDTD(double _length, double _period, unsigned long long _numt);
-	PipeFDTD(double _length, double _period, unsigned long long _numt, double _sonic, double _decay);
+	PipeFDTD(double _length, double _period, size_t _numt);
+	PipeFDTD(double _length, double _period, size_t _numt, double _sonic, double _decay);
 
 protected: /* virtual */
 	bool solve() override;
