+++ AcousticWave/include/WaveFunction.h
@@ -58,7 +58,7 @@ protected: /* parts */
     bool setWaveHeader(unsigned int _srate, unsigned short _sbit);
     bool setWaveData(const WaveData& _data);
 
-    static unsigned long long calcWaveDataSize(double _dura, unsigned int _srate);
+    static size_t calcWaveDataSize(double _dura, unsigned int _srate);
 
 public: /* constructor */
     WaveFunction();
@@ -78,7 +78,7 @@ public: /* public use */
 
     WaveHeader getWaveHeader() const;
     WaveData getWaveData() const;
-    short getWaveData(unsigned long long _tdx) const;
+    short getWaveData(size_t _tdx) const;
 
     static WaveFunction sin(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit);
     static WaveFunction sqr(double _namp, double _freq, double _dura, unsigned int _srate, unsigned short _sbit, double _duty);
