+++ AcousticWave/source/StringFDTD.cpp
@@ -1,13 +1,13 @@
 #include "../include/StringFDTD.h"
 #include "../include/Predefined.h"
 
-StringFDTD::StringFDTD(double _length, double _period, unsigned long long _numt, double _tension, double _density)
+StringFDTD::StringFDTD(double _length, double _period, size_t _numt, double _tension, double _density)
 {
     setBasicCondition(sqrt(_tension / _density), _length, _period, _numt);
     solve();
 }
 
-StringFDTD::StringFDTD(double _length, double _period, unsigned long long _numt, double _tension, double _density, double _decay)
+StringFDTD::StringFDTD(double _length, double _period, size_t _numt, double _tension, double _density, double _decay)
 {
     decay = _decay;
 
@@ -23,9 +23,9 @@ bool StringFDTD::solve()
     setFixedEndCondition(wave[0]);
     setFixedEndCondition(wave[1]);
 
-    for (unsigned long long n = 1; n < numt - 1; ++n)
+    for (size_t n = 1; n < numt - 1; ++n)
     {
-        for (unsigned long long i = 1; i < numx - 1; ++i) 
+        for (size_t i = 1; i < numx - 1; ++i) 
         {
             wave[n + 1][i] = 0;
             wave[n + 1][i] += wave[n][i] * (2.0 + decay * dt);
