+++ AcousticWave/source/WaveFunction.cpp
@@ -57,7 +57,7 @@ WaveFunction WaveFunction::operator*(const WaveFunction& _rhs) const
     ret.header = _rhs.header;
     ret.wdata.resize(wdata.size());
 
-    for (unsigned long long i = 0; i < wdata.size(); ++i)
+    for (size_t i = 0; i < wdata.size(); ++i)
     {
         ret.wdata[i] = wdata[i] * _rhs.wdata[i];
     }
@@ -76,7 +76,7 @@ WaveFunction WaveFunction::operator&(const WaveFunction& _rhs) const
     ret.header = header;
     ret.wdata = wdata;
 
-    for (unsigned long long i = 0; i < _rhs.wdata.size(); ++i)
+    for (size_t i = 0; i < _rhs.wdata.size(); ++i)
     {
         ret.wdata.push_back(_rhs.wdata[i]);
     }
@@ -145,9 +145,9 @@ bool WaveFunction::setWaveData(const WaveData& _data)
     return true;
 }
 
-unsigned long long WaveFunction::calcWaveDataSize(double _dura, unsigned int _srate)
+size_t WaveFunction::calcWaveDataSize(double _dura, unsigned int _srate)
 {
-    return (unsigned long long)(_dura * (double)_srate);
+    return (size_t)(_dura * (double)_srate);
 }
 
 WaveFunction::WaveFunction()
@@ -238,7 +238,7 @@ void WaveFunction::playWave(double* _pch, double _namp)
     }
 
     std::vector<short> vdata = wdata;
-    for (unsigned long long i = 0; i < wdata.size(); ++i)
+    for (size_t i = 0; i < wdata.size(); ++i)
     {
         vdata[i] *= _namp;
     }
@@ -397,7 +397,7 @@ WaveData WaveFunction::getWaveData() const
     return wdata;
 }
 
-short WaveFunction::getWaveData(unsigned long long _tdx) const
+short WaveFunction::getWaveData(size_t _tdx) const
 {
     assert(_tdx < wdata.size());
     return wdata[_tdx];
@@ -411,7 +411,7 @@ WaveFunction WaveFunction::sin(double _namp, double _freq, double _dura, unsigne
     assert(_namp < 1.0);
     double ramp = _namp * pow(2.0, (double)_sbit - 1.0);
     double unit = 1.0 / (double)_srate;
-    unsigned long long size = calcWaveDataSize(_dura, _srate);
+    size_t size = calcWaveDataSize(_dura, _srate);
 
     double tomg = 2.0 * PI * _freq;
 
@@ -435,7 +435,7 @@ WaveFunction WaveFunction::sqr(double _namp, double _freq, double _dura, unsigne
     double ramp = _namp * pow(2.0, (double)_sbit - 1.0);
     double unit = 1.0 / (double)_srate;
     double period = 1.0 / _freq;
-    unsigned long long size = calcWaveDataSize(_dura, _srate);
+    size_t size = calcWaveDataSize(_dura, _srate);
 
     auto step = [period, ramp, _duty](double time)
         {
@@ -465,7 +465,7 @@ WaveFunction WaveFunction::tri(double _namp, double _freq, double _dura, unsigne
     double ramp = _namp * pow(2.0, (double)_sbit - 1.0);
     double unit = 1.0 / (double)_srate;
     double period = 1.0 / _freq;
-    unsigned long long size = calcWaveDataSize(_dura, _srate);
+    size_t size = calcWaveDataSize(_dura, _srate);
 
     auto step = [period, ramp](double time)
         {
@@ -497,7 +497,7 @@ WaveFunction WaveFunction::saw(double _namp, double _freq, double _dura, unsigne
     double ramp = _namp * pow(2.0, (double)_sbit - 1.0);
     double unit = 1.0 / (double)_srate;
     double period = 1.0 / _freq;
-    unsigned long long size = calcWaveDataSize(_dura, _srate);
+    size_t size = calcWaveDataSize(_dura, _srate);
 
     auto step = [period, ramp, _reverse](double time)
         {
@@ -524,7 +524,7 @@ WaveFunction WaveFunction::ofs(double _ofs, double _dura, unsigned int _srate, u
     WaveFunction ret;
     WaveData dat;
 
-    unsigned long long size = calcWaveDataSize(_dura, _srate);
+    size_t size = calcWaveDataSize(_dura, _srate);
 
     for (double i = 0; i < size; ++i)
     {
